Move and rename trial storage-related hooks
