+++ src/components/editor/Bindings/Backfill/BackfillButton.tsx
@@ -2,7 +2,7 @@ import { Box, Stack, Typography } from '@mui/material';
 import BooleanToggleButton from 'components/shared/buttons/BooleanToggleButton';
 import { BooleanString } from 'components/shared/buttons/types';
 import { useEntityWorkflow } from 'context/Workflow';
-import useTrialCollections from 'hooks/useTrialCollections';
+import useTrialCollections from 'hooks/trialStorage/useTrialCollections';
 import { useCallback, useMemo } from 'react';
 import { useIntl } from 'react-intl';
 import {
