+++ src/components/editor/Bindings/UpdateResourceConfigButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@mui/material';
 import { AddCollectionDialogCTAProps } from 'components/shared/Entity/types';
 import invariableStores from 'context/Zustand/invariableStores';
-import useTrialCollections from 'hooks/useTrialCollections';
+import useTrialCollections from 'hooks/trialStorage/useTrialCollections';
 
 import { FormattedMessage } from 'react-intl';
 import {
