+++ src/components/materialization/SourceCapture/AddSourceCaptureToSpecButton.tsx
@@ -1,7 +1,7 @@
 import { Button } from '@mui/material';
 import { AddCollectionDialogCTAProps } from 'components/shared/Entity/types';
 import invariableStores from 'context/Zustand/invariableStores';
-import useTrialCollections from 'hooks/useTrialCollections';
+import useTrialCollections from 'hooks/trialStorage/useTrialCollections';
 import { FormattedMessage } from 'react-intl';
 import { useBinding_prefillResourceConfigs } from 'stores/Binding/hooks';
 import { useBindingStore } from 'stores/Binding/Store';
