+++ src/stores/Binding/Hydrator.tsx
@@ -1,6 +1,6 @@
 import { useEntityType } from 'context/EntityContext';
 import { useEntityWorkflow, useEntityWorkflow_Editing } from 'context/Workflow';
-import useTrialStorageOnly from 'hooks/useTrialStorageOnly';
+import useTrialPrefixes from 'hooks/trialStorage/useTrialPrefixes';
 import { useEffect, useRef } from 'react';
 import { logRocketConsole } from 'services/shared';
 import { useDetailsFormStore } from 'stores/DetailsForm/Store';
@@ -22,7 +22,7 @@ export const BindingHydrator = ({ children }: BaseComponentProps) => {
     const workflow = useEntityWorkflow();
     const editWorkflow = useEntityWorkflow_Editing();
 
-    const getTrialOnlyPrefixes = useTrialStorageOnly();
+    const getTrialPrefixes = useTrialPrefixes();
 
     const connectorTagId = useDetailsFormStore(
         (state) => state.details.data.connectorImage.id
@@ -47,7 +47,7 @@ export const BindingHydrator = ({ children }: BaseComponentProps) => {
                 editWorkflow,
                 entityType,
                 connectorTagId,
-                getTrialOnlyPrefixes,
+                getTrialPrefixes,
                 rehydrating.current
             )
                 .then(
@@ -81,7 +81,7 @@ export const BindingHydrator = ({ children }: BaseComponentProps) => {
         connectorTagId,
         editWorkflow,
         entityType,
-        getTrialOnlyPrefixes,
+        getTrialPrefixes,
         hydrateState,
         setActive,
         setHydrated,
